New: ItemCounter caching like sum, mean, variance, min, max
Mostly for ease-of-use.  Not convinced it is smart, though, because the suggested workflow is
to load data using an ItemCounter and then perform analysis with the CachingItemCounter.  The
(potential) problem is that the two can share data, and mutating the original item counter can
cause the caching item counter to be inconsistent.

The alternative that had been recommended was for the user to handle caching themselves.  This
is potentially safer because the caching mechanism is explicit for the user.